Taught Scala-cask to publish JS model files
