+++ modules/openapi-generator/src/main/java/org/openapitools/codegen/languages/ScalaCaskServerCodegen.java
@@ -134,7 +134,7 @@ public class ScalaCaskServerCodegen extends AbstractScalaCodegen implements Code
 
     @Override
     public String testPackage() {
-        return "src/test/scala";
+        return "jvm/src/test/scala";
     }
 
     public String toModelTestFilename(String name) {
@@ -166,8 +166,8 @@ public class ScalaCaskServerCodegen extends AbstractScalaCodegen implements Code
         modelPackage = ensureProp(CodegenConstants.MODEL_PACKAGE, basePackage + ".model");
 
 
-        final String apiPath = "src/main/scala/" + apiPackage.replace('.', '/');
-        final String modelPath = "src/main/scala/" + modelPackage.replace('.', '/');
+        final String apiPath = "jvm/src/main/scala/" + apiPackage.replace('.', '/');
+        final String modelPath = "shared/src/main/scala/" + modelPackage.replace('.', '/');
 
         final List<String> appFullPath = Arrays.stream(apiPath.split("/")).collect(Collectors.toList());
         final String appFolder = String.join("/", appFullPath.subList(0, appFullPath.size() - 1));
@@ -186,7 +186,7 @@ public class ScalaCaskServerCodegen extends AbstractScalaCodegen implements Code
         supportingFiles.add(new SupportingFile("Dockerfile.mustache", "example", "Dockerfile"));
         supportingFiles.add(new SupportingFile("README.mustache", "", "README.md"));
         supportingFiles.add(new SupportingFile("build.sbt.mustache", "", "build.sbt"));
-        supportingFiles.add(new SupportingFile("bulidAndPublish.yml.mustache", "", ".github/workflows/bulidAndPublish.yml"));
+        supportingFiles.add(new SupportingFile("buildAndPublish.yml.mustache", "", ".github/workflows/buildAndPublish.yml"));
         supportingFiles.add(new SupportingFile("build.sc.mustache", "", "build.sc"));
         supportingFiles.add(new SupportingFile(".scalafmt.conf.mustache", "", ".scalafmt.conf"));
         supportingFiles.add(new SupportingFile("gitignore.mustache", "", ".gitignore"));
@@ -274,6 +274,16 @@ public class ScalaCaskServerCodegen extends AbstractScalaCodegen implements Code
         }
     }
 
+    @Override
+    public String apiFileFolder() {
+        return outputFolder + "/jvm/" + sourceFolder + "/" + apiPackage().replace('.', File.separatorChar);
+    }
+
+    @Override
+    public String modelFileFolder() {
+        return outputFolder + "/shared/" + sourceFolder + "/" + modelPackage().replace('.', File.separatorChar);
+    }
+
     static String capitalise(String p) {
         if (p.length() < 2) {
             return p.toUpperCase(Locale.ROOT);
@@ -323,7 +333,7 @@ public class ScalaCaskServerCodegen extends AbstractScalaCodegen implements Code
         String jsonOpenAPI = SerializerUtils.toJsonString(openAPI);
 
         try {
-            String outputFile = getOutputDir() + "/" + getResourceFolder() + "/openapi.json";
+            String outputFile = getOutputDir() + "/jvm/" + getResourceFolder() + "/openapi.json";
             FileUtils.writeStringToFile(new File(outputFile), jsonOpenAPI, StandardCharsets.UTF_8);
             LOGGER.info("wrote file to {}", outputFile);
         } catch (Exception e) {
