+++ modules/openapi-generator/src/main/resources/scala-cask/build.sc.mustache
@@ -28,8 +28,8 @@ object {{artifactId}} extends SbtModule with ScalafmtModule with PublishModule {
     ivy"com.lihaoyi::upickle:3.2.0"
   )
 
-  override def sources = T.sources(millSourcePath / os.up / "src" / "main" / "scala")
-  override def resources = T.sources(millSourcePath / os.up / "src" / "main" / "resources")
+  override def sources = T.sources(millSourcePath / os.up / "shared" / "src" / "main" / "scala")
+  override def resources = T.sources(millSourcePath / os.up / "shared" / "src" / "main" / "resources")
 
   object test extends SbtModuleTests {
     def ivyDeps = Agg(
