+++ modules/openapi-generator/src/main/resources/scala-cask/buildAndPublish.yml.mustache
@@ -30,10 +30,7 @@ jobs:
           ${{{openbrackets}}} runner.os {{{closebrackets}}}-sbt-
 
     - name: Build with sbt
-      run: sbt clean compile
-
-    - name: Test with sbt
-      run: sbt test
+      run: sbt clean compile test
 
     - name: Publish to GitHub Packages
       run: sbt publish
