+++ samples/server/petstore/scala-cask/.github/workflows/buildAndPublish.yml
@@ -30,10 +30,7 @@ jobs:
           ${{ runner.os }}-sbt-
 
     - name: Build with sbt
-      run: sbt clean compile
-
-    - name: Test with sbt
-      run: sbt test
+      run: sbt clean compile test
 
     - name: Publish to GitHub Packages
       run: sbt publish
