+++ samples/server/petstore/scala-cask/.openapi-generator/FILES
@@ -1,5 +1,6 @@
-.github/workflows/bulidAndPublish.yml
+.github/workflows/buildAndPublish.yml
 .gitignore
+.openapi-generator-ignore
 .scalafmt.conf
 README.md
 README.md
@@ -7,33 +8,39 @@ build.sbt
 build.sc
 example/Dockerfile
 example/Server.scala
+jvm/src/main/scala/sample/cask/AppRoutes.scala
+jvm/src/main/scala/sample/cask/BaseApp.scala
+jvm/src/main/scala/sample/cask/ExampleApp.scala
+jvm/src/main/scala/sample/cask/api/OpenApiRoutes.scala
+jvm/src/main/scala/sample/cask/api/PetRoutes.scala
+jvm/src/main/scala/sample/cask/api/PetRoutes.scala
+jvm/src/main/scala/sample/cask/api/PetService.scala
+jvm/src/main/scala/sample/cask/api/StoreRoutes.scala
+jvm/src/main/scala/sample/cask/api/StoreRoutes.scala
+jvm/src/main/scala/sample/cask/api/StoreService.scala
+jvm/src/main/scala/sample/cask/api/UserRoutes.scala
+jvm/src/main/scala/sample/cask/api/UserRoutes.scala
+jvm/src/main/scala/sample/cask/api/UserService.scala
+jvm/src/main/scala/sample/cask/api/package.scala
+jvm/src/main/scala/sample/cask/package.scala
+jvm/src/test/scala/sample/cask/model/ApiResponseTest.scala
+jvm/src/test/scala/sample/cask/model/CategoryTest.scala
+jvm/src/test/scala/sample/cask/model/OrderTest.scala
+jvm/src/test/scala/sample/cask/model/PetTest.scala
+jvm/src/test/scala/sample/cask/model/TagTest.scala
+jvm/src/test/scala/sample/cask/model/UserTest.scala
 project/build.properties
 project/plugins.sbt
-src/main/scala/sample/cask/AppRoutes.scala
-src/main/scala/sample/cask/BaseApp.scala
-src/main/scala/sample/cask/ExampleApp.scala
-src/main/scala/sample/cask/api/OpenApiRoutes.scala
-src/main/scala/sample/cask/api/PetRoutes.scala
-src/main/scala/sample/cask/api/PetRoutes.scala
-src/main/scala/sample/cask/api/PetService.scala
-src/main/scala/sample/cask/api/StoreRoutes.scala
-src/main/scala/sample/cask/api/StoreRoutes.scala
-src/main/scala/sample/cask/api/StoreService.scala
-src/main/scala/sample/cask/api/UserRoutes.scala
-src/main/scala/sample/cask/api/UserRoutes.scala
-src/main/scala/sample/cask/api/UserService.scala
-src/main/scala/sample/cask/api/package.scala
-src/main/scala/sample/cask/model/ApiResponse.scala
-src/main/scala/sample/cask/model/ApiResponseData.scala
-src/main/scala/sample/cask/model/Category.scala
-src/main/scala/sample/cask/model/CategoryData.scala
-src/main/scala/sample/cask/model/Order.scala
-src/main/scala/sample/cask/model/OrderData.scala
-src/main/scala/sample/cask/model/Pet.scala
-src/main/scala/sample/cask/model/PetData.scala
-src/main/scala/sample/cask/model/Tag.scala
-src/main/scala/sample/cask/model/TagData.scala
-src/main/scala/sample/cask/model/User.scala
-src/main/scala/sample/cask/model/UserData.scala
-src/main/scala/sample/cask/model/package.scala
-src/main/scala/sample/cask/package.scala
+shared/src/main/scala/sample/cask/model/ApiResponse.scala
+shared/src/main/scala/sample/cask/model/ApiResponseData.scala
+shared/src/main/scala/sample/cask/model/Category.scala
+shared/src/main/scala/sample/cask/model/CategoryData.scala
+shared/src/main/scala/sample/cask/model/Order.scala
+shared/src/main/scala/sample/cask/model/OrderData.scala
+shared/src/main/scala/sample/cask/model/Pet.scala
+shared/src/main/scala/sample/cask/model/PetData.scala
+shared/src/main/scala/sample/cask/model/Tag.scala
+shared/src/main/scala/sample/cask/model/TagData.scala
+shared/src/main/scala/sample/cask/model/User.scala
+shared/src/main/scala/sample/cask/model/UserData.scala
+shared/src/main/scala/sample/cask/model/package.scala
