+++ samples/server/petstore/scala-cask/build.sbt
@@ -1,23 +1,45 @@
-name := "scala-cask-petstore"
-organization:="cask.groupId"
-version := "0.0.1-SNAPSHOT"
-scalaVersion := "3.3.1"
-scalafmtOnCompile := true
-libraryDependencies ++= Seq(
-  "com.lihaoyi" %% "cask" % "0.9.2" ,
-  "com.lihaoyi" %% "upickle" % "3.2.0",
-  "org.scalatest" %% "scalatest" % "3.2.18" % Test
+ThisBuild / name := "scala-cask-petstore"
+ThisBuild / organization := "cask.groupId"
+ThisBuild / version := "0.0.1-SNAPSHOT"
+ThisBuild / scalaVersion := "3.4.1"
+ThisBuild / scalafmtOnCompile := true
+
+// Common settings
+lazy val commonSettings = Seq(
+  libraryDependencies ++= Seq(
+    "com.lihaoyi" %%% "upickle" % "3.2.0",
+    "org.scalatest" %%% "scalatest" % "3.2.18" % Test
+  )
 )
 
-publishMavenStyle := true
+
+lazy val app = crossProject(JSPlatform, JVMPlatform).in(file(".")).
+  settings(commonSettings).
+  jvmSettings(
+    libraryDependencies += "com.lihaoyi" %% "cask" % "0.9.2"
+  ).
+  jsSettings(
+    scalaJSUseMainModuleInitializer := true,
+  )
+  
+
+lazy val root = project.in(file(".")).
+  aggregate(app.js, app.jvm).
+  settings(
+    publish := {},
+    publishLocal := {},
+  )
+
+
+ThisBuild / publishMavenStyle := true
 
 val githubUser = "GIT_USER_ID"
 val githubRepo = "GIT_REPO_ID"
-publishTo := Some("GitHub Package Registry" at s"https://maven.pkg.github.com/$githubUser/$githubRepo")
+ThisBuild / publishTo := Some("GitHub Package Registry" at s"https://maven.pkg.github.com/$githubUser/$githubRepo")
 
 sys.env.get("GITHUB_TOKEN") match {
   case Some(token) if !token.isEmpty =>
-    credentials += Credentials(
+    ThisBuild / credentials += Credentials(
       "GitHub Package Registry",
       "maven.pkg.github.com",
       githubUser,
