+++ samples/server/petstore/scala-cask/jvm/src/main/resources/openapi.json
