+++ samples/server/petstore/scala-cask/jvm/src/main/scala/sample/cask/api/UserService.scala
