+++ samples/server/petstore/scala-cask/jvm/src/test/scala/sample/cask/model/CategoryTest.scala
@@ -3,10 +3,12 @@
  * This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.
  *
  * OpenAPI spec version: 1.0.0
+ *
  * Contact: team@openapitools.org
  *
- * NOTE: This class is auto generated by the swagger code generator program.
- * https://github.com/swagger-api/swagger-codegen.git
+ * NOTE: This class is auto generated by OpenAPI Generator.
+ *
+ * https://openapi-generator.tech
  */
 
 // this model was generated using modelTest.mustache
@@ -16,15 +18,15 @@ import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import scala.util.*
 
-class TagTest extends AnyWordSpec with Matchers {
+class CategoryTest extends AnyWordSpec with Matchers {
 
-    "Tag.fromJson" should {
+    "Category.fromJson" should {
         """not parse invalid json""" in {
-           val Failure(err) = Try(TagData.fromJsonString("invalid jason"))
+           val Failure(err) = Try(CategoryData.fromJsonString("invalid jason"))
            err.getMessage should startWith ("Error parsing json 'invalid jason'")
         }
         """parse """ ignore {
-           val Failure(err : ValidationErrors) = TagData.fromJsonString("""""").validated()
+           val Failure(err : ValidationErrors) = CategoryData.fromJsonString("""""").validated()
 
            sys.error("TODO")
         }
