+++ samples/server/petstore/scala-cask/jvm/src/test/scala/sample/cask/model/OrderTest.scala
@@ -3,28 +3,31 @@
  * This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.
  *
  * OpenAPI spec version: 1.0.0
+ *
  * Contact: team@openapitools.org
  *
- * NOTE: This class is auto generated by the swagger code generator program.
- * https://github.com/swagger-api/swagger-codegen.git
+ * NOTE: This class is auto generated by OpenAPI Generator.
+ *
+ * https://openapi-generator.tech
  */
 
 // this model was generated using modelTest.mustache
 package sample.cask.model
+import java.time.OffsetDateTime
 
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import scala.util.*
 
-class ApiResponseTest extends AnyWordSpec with Matchers {
+class OrderTest extends AnyWordSpec with Matchers {
 
-    "ApiResponse.fromJson" should {
+    "Order.fromJson" should {
         """not parse invalid json""" in {
-           val Failure(err) = Try(ApiResponseData.fromJsonString("invalid jason"))
+           val Failure(err) = Try(OrderData.fromJsonString("invalid jason"))
            err.getMessage should startWith ("Error parsing json 'invalid jason'")
         }
         """parse """ ignore {
-           val Failure(err : ValidationErrors) = ApiResponseData.fromJsonString("""""").validated()
+           val Failure(err : ValidationErrors) = OrderData.fromJsonString("""""").validated()
 
            sys.error("TODO")
         }
