+++ samples/server/petstore/scala-cask/jvm/src/test/scala/sample/cask/model/PetTest.scala
@@ -3,28 +3,32 @@
  * This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.
  *
  * OpenAPI spec version: 1.0.0
+ *
  * Contact: team@openapitools.org
  *
- * NOTE: This class is auto generated by the swagger code generator program.
- * https://github.com/swagger-api/swagger-codegen.git
+ * NOTE: This class is auto generated by OpenAPI Generator.
+ *
+ * https://openapi-generator.tech
  */
 
 // this model was generated using modelTest.mustache
 package sample.cask.model
+import sample.cask.model.Category
+import sample.cask.model.Tag
 
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import scala.util.*
 
-class UserTest extends AnyWordSpec with Matchers {
+class PetTest extends AnyWordSpec with Matchers {
 
-    "User.fromJson" should {
+    "Pet.fromJson" should {
         """not parse invalid json""" in {
-           val Failure(err) = Try(UserData.fromJsonString("invalid jason"))
+           val Failure(err) = Try(PetData.fromJsonString("invalid jason"))
            err.getMessage should startWith ("Error parsing json 'invalid jason'")
         }
         """parse """ ignore {
-           val Failure(err : ValidationErrors) = UserData.fromJsonString("""""").validated()
+           val Failure(err : ValidationErrors) = PetData.fromJsonString("""""").validated()
 
            sys.error("TODO")
         }
