+++ samples/server/petstore/scala-cask/jvm/src/test/scala/sample/cask/model/UserTest.scala
@@ -3,10 +3,12 @@
  * This is a sample server Petstore server. For this sample, you can use the api key `special-key` to test the authorization filters.
  *
  * OpenAPI spec version: 1.0.0
+ *
  * Contact: team@openapitools.org
  *
- * NOTE: This class is auto generated by the swagger code generator program.
- * https://github.com/swagger-api/swagger-codegen.git
+ * NOTE: This class is auto generated by OpenAPI Generator.
+ *
+ * https://openapi-generator.tech
  */
 
 // this model was generated using modelTest.mustache
@@ -16,15 +18,15 @@ import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import scala.util.*
 
-class CategoryTest extends AnyWordSpec with Matchers {
+class UserTest extends AnyWordSpec with Matchers {
 
-    "Category.fromJson" should {
+    "User.fromJson" should {
         """not parse invalid json""" in {
-           val Failure(err) = Try(CategoryData.fromJsonString("invalid jason"))
+           val Failure(err) = Try(UserData.fromJsonString("invalid jason"))
            err.getMessage should startWith ("Error parsing json 'invalid jason'")
         }
         """parse """ ignore {
-           val Failure(err : ValidationErrors) = CategoryData.fromJsonString("""""").validated()
+           val Failure(err : ValidationErrors) = UserData.fromJsonString("""""").validated()
 
            sys.error("TODO")
         }
