+++ samples/server/petstore/scala-cask/shared/src/main/scala/sample/cask/model/Category.scala
