+++ samples/server/petstore/scala-cask/shared/src/main/scala/sample/cask/model/TagData.scala
